response data refactor - source top 15 data is ready
